requirements.txt
pandas
python-dotenv
tabulate
tushare
